+++ src/main/java/simple/myboard/myprac/service/MemberService.java
@@ -6,10 +6,10 @@ import simple.myboard.myprac.domain.Member;
 public interface MemberService {
 
     public void setMemberDao(MemberDao memberDao);
-    public abstract Member getMemberBySeq(int memberSeq);
+    public abstract Member getMemberBySeq(Long memberSeq);
     public abstract void addMember(Member member);
     public abstract void updateMember(Member member);
-    public abstract void deleteMemberBySeq(int memberSeq);
-    public abstract int getLastIndexMember();
+    public abstract void deleteMemberBySeq(Long memberSeq);
+    public abstract Long getLastIndexMember();
     
 }
